+++ packages/plugin-nm/sources/NodeModulesLinker.ts
@@ -61,8 +61,11 @@ export class NodeModulesLinker implements Linker {
       throw err;
     }
 
-    const startingCwd = opts.project.configuration.startingCwd;
-    return locatorInfo.locations.find(location => ppath.contains(startingCwd, location)) || locatorInfo.locations[0];
+    // Sort locations from shallowest to deepest in terms of directory nesting
+    const sortedLocations = locatorInfo.locations.sort((loc1, loc2) => loc1.split(ppath.sep).length - loc2.split(ppath.sep).length);
+    // Find the location with shallowest directory nesting that starts inside node_modules of cwd
+    const startingCwdModules = ppath.join(opts.project.configuration.startingCwd, NODE_MODULES);
+    return sortedLocations.find(location => ppath.contains(startingCwdModules, location)) || locatorInfo.locations[0];
   }
 
   async findPackageLocator(location: PortablePath, opts: LinkOptions) {